% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_algorithms.R
\name{get_available_algorithms}
\alias{get_available_algorithms}
\title{Function for getting available grouping algorithms}
\usage{
get_available_algorithms()
}
\value{
vector
}
\description{
Returns vector of names of available algorithms
}
