% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_required_parameters.R
\name{get_required_parameters}
\alias{get_required_parameters}
\title{Function for returning required parameters of a grouping algorithm}
\usage{
get_required_parameters(algorithm)
}
\arguments{
\item{algorithm}{Character name of the algorithm (same name as shown in get_available_algorithms)}
}
\value{
list
}
\description{
Returns required parameters list
}
