% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group}
\alias{group}
\title{Perform grouping.}
\usage{
group(model, copy = FALSE)
}
\arguments{
\item{model}{one model from group.formation.algorithms package}

\item{copy}{TRUE/FALSE make a copy of the model and perform the grouping on copy}
}
\value{
model with the groups created
}
\description{
Function that executes the grouping over the provided algorithm object.
}
\examples{
\dontrun{
g <- GroupAlgorithm$new(
       data.frame(a = c(1,2), b = c(2,2)),
       list(c = 1, d = "a"),
       c(3,4))
g2 <- group(g)
}

}
