% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping_function.R
\name{group2}
\alias{group2}
\title{Groups counts based on range.}
\usage{
group2(x, group_of, FUN = "sum", age_labels = F)
}
\arguments{
\item{x}{A dataframe}

\item{group_of}{The range to group together.}

\item{FUN}{A string of the function to be performed in the aggregate section.
Default = "sum"}

\item{age_labels}{Logical. If true, age labels are attempted. Only works if
1st age group starts at 0.}
}
\value{
A dataset that has been groupped by \code{group_of} and has had
  \code{FUN} perfomed on the grouping.
}
\description{
Created to count the number of people in a certain age range. Function tested
using a data frame with continuous age values in one column, and counts of
people in another column.
}
\examples{
### Sum the number of men in each user defined age group
 # create data
 x <- data.frame(age = c(1:6), males = c(0,1,0,2,3,5), females = c(4,6,2,1,0,1))
 group2(x, 3, "sum")
}

