% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupAlgorithm.R
\name{group_algorithm}
\alias{group_algorithm}
\title{Constructor function for GroupAlgorithm R6 class.}
\usage{
group_algorithm(features, preferences = NULL, parameters = NULL, bounds)
}
\arguments{
\item{features}{data.frame containing the grouping data}

\item{parameters}{list containing algorithm parameters}

\item{bounds}{num scalar or vector of 2 elements with bounds}

\item{preferencds}{matrix of preferences [NOT USED]}
}
\value{
GroupAlgorithm object.
}
\description{
Creates GroupAlgorithm object.
}
