% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupAlgorithm.R
\name{GroupAlgorithm}
\alias{GroupAlgorithm}
\title{R6 Class GroupAlgorithm}
\description{
GroupAlgorithm is basic (maternal) R6 class for creating student groupings.
}
\details{
GroupAlgorithm is basic (maternal) R6 class for creating student groupings. It
does not implement any grouping algorithm, but it implements helper
functionalities and proper behavior. This class should be used as a template
for creating new Algorithm class inheriting from this one.

All methods constructors should be ending with "_algorithm".
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{features}}{data.frame containing the grouping data}

\item{\code{preferences}}{num matrix containing the preferences data}

\item{\code{parameters}}{list containing algorithm parameters}

\item{\code{bounds}}{num scalar or vector of 2 elements with bounds}

\item{\code{result}}{vector containing assignment of each student to group}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GroupAlgorithm-new}{\code{GroupAlgorithm$new()}}
\item \href{#method-GroupAlgorithm-print}{\code{GroupAlgorithm$print()}}
\item \href{#method-GroupAlgorithm-get_required_parameters}{\code{GroupAlgorithm$get_required_parameters()}}
\item \href{#method-GroupAlgorithm-group}{\code{GroupAlgorithm$group()}}
\item \href{#method-GroupAlgorithm-clone}{\code{GroupAlgorithm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupAlgorithm-new"></a>}}
\if{latex}{\out{\hypertarget{method-GroupAlgorithm-new}{}}}
\subsection{Method \code{new()}}{
Create new object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupAlgorithm$new(
  features,
  preferences,
  parameters = NULL,
  param_requirements,
  bounds
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{features}}{data.frame containing grouping data}

\item{\code{parameters}}{list containing algorithm configuration}

\item{\code{param_requirements}}{list containing all required parameters and possible values}

\item{\code{bounds}}{num scalar or vector of 2 elements with group bounds}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupAlgorithm-print"></a>}}
\if{latex}{\out{\hypertarget{method-GroupAlgorithm-print}{}}}
\subsection{Method \code{print()}}{
Print the information about the algorithm.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupAlgorithm$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{dots for compatibility reasons}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupAlgorithm-get_required_parameters"></a>}}
\if{latex}{\out{\hypertarget{method-GroupAlgorithm-get_required_parameters}{}}}
\subsection{Method \code{get_required_parameters()}}{
Prints out required parameters and their specification
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupAlgorithm$get_required_parameters()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupAlgorithm-group"></a>}}
\if{latex}{\out{\hypertarget{method-GroupAlgorithm-group}{}}}
\subsection{Method \code{group()}}{
Perform grouping of students.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupAlgorithm$group()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupAlgorithm-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GroupAlgorithm-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupAlgorithm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
