% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping.R
\docType{package}
\name{grouping-package}
\alias{grouping}
\alias{grouping-package}
\title{grouping: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
