% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KardanAlgorithm.R
\name{kardan_algorithm}
\alias{kardan_algorithm}
\title{Constructor function for KardanAlgorithm R6 class.}
\usage{
kardan_algorithm(features, preferences = NULL, parameters, bounds)
}
\arguments{
\item{features}{data.frame containing the grouping data}

\item{preferences}{matrix containing student preferences}

\item{parameters}{list containing algorithm parameters}

\item{bounds}{num scalar or vector of 2 elements with bounds}
}
\value{
KardanAlgorithm object.
}
\description{
Creates KardanAlgorithm object.
}
