% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomAlgorithm.R
\name{random_algorithm}
\alias{random_algorithm}
\title{Constructor function for RandomAlgorithm R6 class.}
\usage{
random_algorithm(features, parameters, bounds)
}
\arguments{
\item{features}{data.frame containing the grouping data}

\item{parameters}{list containing algorithm parameters}

\item{bounds}{num scalar or vector of 2 elements with bounds}
}
\value{
RandomAlgorithm object.
}
\description{
Creates RandomAlgorithm object.
}
