% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomAlgorithm.R
\name{RandomAlgorithm}
\alias{RandomAlgorithm}
\title{R6 Class RandomAlgorithm}
\description{
RandomAlgorithm is R6 class for creating student random groupings.
}
\details{
RandomAlgorithm is very simple (example) algorithm for creating the groups.
}
\section{Super class}{
\code{\link[grouping:GroupAlgorithm]{grouping::GroupAlgorithm}} -> \code{RandomAlgorithm}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RandomAlgorithm-group}{\code{RandomAlgorithm$group()}}
\item \href{#method-RandomAlgorithm-clone}{\code{RandomAlgorithm$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="grouping" data-topic="GroupAlgorithm" data-id="get_required_parameters"><a href='../../grouping/html/GroupAlgorithm.html#method-GroupAlgorithm-get_required_parameters'><code>grouping::GroupAlgorithm$get_required_parameters()</code></a></span></li>
<li><span class="pkg-link" data-pkg="grouping" data-topic="GroupAlgorithm" data-id="initialize"><a href='../../grouping/html/GroupAlgorithm.html#method-GroupAlgorithm-initialize'><code>grouping::GroupAlgorithm$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="grouping" data-topic="GroupAlgorithm" data-id="print"><a href='../../grouping/html/GroupAlgorithm.html#method-GroupAlgorithm-print'><code>grouping::GroupAlgorithm$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RandomAlgorithm-group"></a>}}
\if{latex}{\out{\hypertarget{method-RandomAlgorithm-group}{}}}
\subsection{Method \code{group()}}{
Perform grouping of students.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandomAlgorithm$group()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RandomAlgorithm-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RandomAlgorithm-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandomAlgorithm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
