% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KardanAlgorithm.R
\name{KardanAlgorithm}
\alias{KardanAlgorithm}
\title{R6 Class KardanAlgorithm}
\description{
KardanAlgorithm is R6 class for creating student random groupings.
}
\details{
KardanAlgorithm is the grouping model developed by Kardan et. al. based on homogeneous
and heterogeneous attributes, student preferences. The model computes compatibility
measure and then performs the mathematical modeling (long-step simplex optimization).
Fairness is guaranteed by initial step, which checks the convergence to solution.
The algorithm is not suitable for large groups of students (> 100). The algorithm
is implemented without fairness step.

Ahmad A. Kardan & Hamid Sadeghi (2016).
\emph{An efficacious dynamic mathematical modelling approach for creation
 of best collaborative groups}. Mathematical and Computer Modelling
 of Dynamical Systems, 22:1, 39-53, DOI: \url{https://doi.org/10.1080/13873954.2015.1086382}
}
\section{Super class}{
\code{\link[grouping:GroupAlgorithm]{grouping::GroupAlgorithm}} -> \code{KardanAlgorithm}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{preferences}}{num matrix containing the preferences data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KardanAlgorithm-new}{\code{KardanAlgorithm$new()}}
\item \href{#method-KardanAlgorithm-group}{\code{KardanAlgorithm$group()}}
\item \href{#method-KardanAlgorithm-clone}{\code{KardanAlgorithm$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="grouping" data-topic="GroupAlgorithm" data-id="get_required_parameters"><a href='../../grouping/html/GroupAlgorithm.html#method-GroupAlgorithm-get_required_parameters'><code>grouping::GroupAlgorithm$get_required_parameters()</code></a></span></li>
<li><span class="pkg-link" data-pkg="grouping" data-topic="GroupAlgorithm" data-id="print"><a href='../../grouping/html/GroupAlgorithm.html#method-GroupAlgorithm-print'><code>grouping::GroupAlgorithm$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KardanAlgorithm-new"></a>}}
\if{latex}{\out{\hypertarget{method-KardanAlgorithm-new}{}}}
\subsection{Method \code{new()}}{
Create new object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KardanAlgorithm$new(
  features,
  preferences,
  parameters = NULL,
  param_requirements,
  bounds
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{features}}{data.frame containing grouping data}

\item{\code{preferences}}{matrix containing student preferences}

\item{\code{parameters}}{list containing algorithm configuration}

\item{\code{param_requirements}}{list containing all required parameters and possible values}

\item{\code{bounds}}{num scalar or vector of 2 elements with group bounds}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KardanAlgorithm-group"></a>}}
\if{latex}{\out{\hypertarget{method-KardanAlgorithm-group}{}}}
\subsection{Method \code{group()}}{
Perform grouping of students.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KardanAlgorithm$group()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KardanAlgorithm-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KardanAlgorithm-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KardanAlgorithm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
